+++ blocks/reaction/reaction.js
@@ -5,7 +5,6 @@
 import { htmlToElement, apiRequest } from '../../scripts/scripts.js';
 import { isSignedInUser, getUserData } from '../../scripts/profile.js';
 
-let reaction = {};
 let userDetails = {};
 let isSignedIn = false;
 const config = {};
@@ -14,22 +13,19 @@ const config = {};
  * Submits or fetches the reaction based on the current state of `reaction`
  * @returns {Promise<void>}
  */
-async function handleReaction() {
+async function handleReaction(reaction = null) {
   const postData = {
     story_id: config.storyId,
     user_id: userDetails.id,
     ...(reaction && { reaction }),
   };
-  const endpoint = reaction ? '/submitReaction' : '/getReactions';
+  const endpoint = reaction ? '/postReaction' : '/getReactions';
   try {
     const { payload } = await apiRequest({
       method: 'POST',
       endpoint,
       data: postData,
     });
-    if (!reaction) {
-      reaction = payload?.reaction_name || '';
-    }
     console.log(reaction ? 'Reaction submitted' : 'Reaction fetched:', reaction || payload);
   } catch (error) {
     console.error('Error handling reaction:', error);
@@ -42,11 +38,15 @@ async function handleReaction() {
  * @returns {Promise<void>}
  */
 async function initReaction(block) {
+  if (isSignedIn) {
+    userDetails = await getUserData();
+  }
+
   const reactionIcons = block.querySelectorAll('.reaction-container div:nth-child(2) > div p');
-  const userReaction = userDetails.id ? reaction.reaction_name : '';
+  const userReaction = userDetails.id ? await handleReaction() : '';
   reactionIcons.forEach((reactionIcon) => {
-    const selectedReaction = reactionIcon.querySelector('img').getAttribute('data-icon-name');
-    if (selectedReaction === userReaction) {
+    const reaction = reactionIcon.querySelector('img').getAttribute('data-icon-name');
+    if (reaction === userReaction) {
       reactionIcon.classList.add('active');
     }
 
@@ -55,7 +55,7 @@ async function initReaction(block) {
       if (isSignedIn) {
         // Activate the clicked icon
         reactionIcon.classList.add('active');
-        await handleReaction(selectedReaction);
+        await handleReaction(reaction);
       } else {
         const reactionContainer = block.querySelector('.reaction-container div:nth-child(2)');
         if (!reactionContainer.querySelector('.info')) {
@@ -77,13 +77,7 @@ async function initReaction(block) {
  * @param {Element} block - The reaction block element
  */
 export default async function decorate(block) {
-  isSignedIn = await isSignedInUser();
-  if (!isSignedIn) {
-    initReaction(block);
-    return;
-  }
-  userDetails = await getUserData() || [];
-  reaction = await handleReaction();
   config.storyId = document.querySelector('meta[name="storyid"]')?.content;
+  isSignedIn = await isSignedInUser();
   initReaction(block);
 }
